# src/docs/apispec.yml
openapi: 3.0.0
info:
  title: Simulation API
  description: API for create and monitor Agent Based Simulations @ Göteborgsregionen
  version: 1.0.0
servers:
  - url: https://api.innovationsarenan.dev/sim
    description: Main production server
security:
  - ApiKeyAuth: []
paths:
...
  /simulations:
    post:
      summary: Create a new simulation
      description: Creates a new simulation with the specified parameters and generates associated agents
      operationId: createSimulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agentCount
                - name
              properties:
                agentCount:
                  type: integer
                  description: Number of agents to create for the simulation
                  example: 5
                name:
                  type: string
                  description: Name of the simulation
                  example: "Climate Change Discussion"
                description:
                  type: string
                  description: Optional description of the simulation
                  example: "A discussion about climate change policies"
                topic:
                  type: string
                  description: Topic for the simulation discussion
                  example: "Climate Change"
                environment:
                  $ref: "#/components/schemas/Environment"
      responses:
        "201":
          description: Simulation created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  simulation:
                    $ref: "#/components/schemas/Simulation"
                  agents:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "uC3BVPe3"
                        name:
                          type: string
                          example: "Anna Andersson"
                        sex:
                          type: string
                          enum: [male, female]
                          example: "female"
        "400":
          description: Invalid input
        "500":
          description: Server error
  /simulations/{simulationId}/start:
    patch:
      summary: Start a simulation
      description: Changes the state of a simulation from 'primed' to 'running' and initiates conversations between agents
      operationId: startSimulation
      parameters:
        - name: simulationId
          in: path
          required: true
          description: ID of the simulation to start
          schema:
            type: string
            example: "iLVieMGDrI6GTBLq"
      responses:
        "200":
          description: Simulation started successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Simulation"
        "400":
          description: Missing simulation ID or invalid input
        "404":
          description: Simulation not found
        "500":
          description: Server error
  /simulations/{simulationId}/stop:
    patch:
      summary: Stop a simulation
      description: Changes the state of a simulation from 'running' to 'stopped'
      operationId: stopSimulation
      parameters:
        - name: simulationId
          in: path
          required: true
          description: ID of the simulation to stop
          schema:
            type: string
            example: "iLVieMGDrI6GTBLq"
      responses:
        "200":
          description: Simulation stopped successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Simulation"
        "400":
          description: Missing simulation ID or invalid input
        "404":
          description: Simulation not found
        "500":
          description: Server error
  /conversations:
    post:
      summary: Create a new conversation
      description: Creates a new conversation between two agents in a simulation
      operationId: createConversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - senderId
                - recieverId
                - simulationId
              properties:
                senderId:
                  type: string
                  description: ID of the agent initiating the conversation
                  example: "uC3BVPe3"
                recieverId:
                  type: string
                  description: ID of the agent receiving the conversation
                  example: "bH8KlMn2"
                simulationId:
                  type: string
                  description: ID of the simulation this conversation belongs to
                  example: "iLVieMGDrI6GTBLq"
      responses:
        "201":
          description: Conversation created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "400":
          description: Invalid input
        "500":
          description: Server error
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: authorization
      description: API key authentication using Bearer token
  schemas:
...
    Tokens:
      type: object
      required:
        - promptTokens
        - completionTokens
      properties:
        promptTokens:
          type: integer
          example: 234
        completionTokens:
          type: integer
          example: 544
    Stats:
      type: object
      properties:
        messages:
          type: integer
          example: 12
        conversations:
          type: integer
          example: 3
        tokens:
          $ref: "#/components/schemas/Tokens"
...
    Simulation:
      type: object
      required:
        - id
        - agentCount
        - state
        - type
        - name
        - topic
      properties:
        id:
          type: string
          description: Unique identifier for the simulation
          example: "iLVieMGDrI6GTBLq"
        agentCount:
          type: integer
          description: Number of agents in the simulation
          example: 5
        state:
          type: string
          enum: [primed, running, ended, stopped]
          description: Current state of the simulation
          example: "primed"
        type:
          type: string
          enum: [discussion, attitude]
          description: Type of simulation
          example: "discussion"
        name:
          type: string
          description: Name of the simulation
          example: "Climate Change Discussion"
        description:
          type: string
          description: Optional description of the simulation
          example: "A discussion about climate change policies"
        environment:
          $ref: "#/components/schemas/Environment"
        topic:
          type: string
          description: Topic for the simulation
          example: "Climate Change"
        stats:
          $ref: "#/components/schemas/Stats"
    Environment:
      type: object
      required:
        - id
        - description
        - objectives
        - constraints
        - values
      properties:
        id:
          type: string
          description: Unique identifier for the environment
          example: "env123"
        description:
          type: string
          description: Description of the environment
          example: "A formal meeting room setting"
        objectives:
          type: array
          description: Goals for the environment
          items:
            type: string
            example: "Reach consensus on climate policy"
        constraints:
          type: array
          description: Constraints within the environment
          items:
            type: string
            example: "Budget limitations"
        values:
          type: array
          description: Values emphasized in the environment
          items:
            type: string
            example: "Sustainability"
    Conversation:
      type: object
      required:
        - id
        - simulationId
        - topic
        - active
        - participants
        - messages
      properties:
        id:
          type: string
          description: Unique identifier for the conversation
          example: "conv123456"
        simulationId:
          type: string
          description: ID of the simulation this conversation belongs to
          example: "iLVieMGDrI6GTBLq"
        topic:
          type: string
          description: Topic of the conversation
          example: "Climate Change"
        active:
          type: boolean
          description: Whether the conversation is currently active
          example: true
        participants:
          type: array
          description: IDs of agents participating in the conversation
          items:
            type: string
            example: "uC3BVPe3"
        messages:
          type: array
          description: Messages in the conversation
          items:
            $ref: "#/components/schemas/Message"
        stats:
          $ref: "#/components/schemas/Stats"
    Message:
      type: object
      required:
        - parentId
        - senderId
        - content
        - simulationId
        - tokens
      properties:
        parentId:
          type: string
          description: ID of the conversation this message belongs to
          example: "conv123456"
        senderId:
          type: string
          description: ID of the agent who sent the message
          example: "uC3BVPe3"
        content:
          type: string
          description: Content of the message
          example: "I think we should consider renewable energy sources."
        simulationId:
          type: string
          description: ID of the simulation this message belongs to
          example: "iLVieMGDrI6GTBLq"
        tokens:
          $ref: "#/components/schemas/Tokens"
...